+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.2.2
 
-* Now default to `vknabel/BashArrow`
+* Now default to `vknabel/BashArrow`. See [vknabel/Archery#0.2.1](https://github.com/vknabel/Archery/blob/master/CHANGELOG.md#021)
 
 ## 0.2.1